+++ src/component/ViewModel.js
@@ -17,7 +17,8 @@ ViewModel.prototype = {
 		var expObj = lexer(path);
 		var evalStr = Renderer.getExpEvalStr(this,expObj);
 		if(arguments.length > 1){
-			eval(evalStr + '= "'+ val.replace(/\n/mg,'\\n') +'"');
+			//fix \r\n on IE8
+			eval(evalStr + '= "'+ val.replace(/\r\n|\n/mg,'\\n') +'"');
 			return this;
 		}else{
 			return eval(evalStr);